+++ android/src/main/kotlin/com/maktubcompany/easy_blue_printer/plugin/domain/repository/BluetoothRepository.kt
@@ -11,4 +11,5 @@ interface BluetoothRepository {
     fun isConnected(): Boolean
     fun printImage(data: ByteArray, align: Int): Boolean
     fun configurePrinter(paperWidth: Int)
+    fun commitPrint(): Boolean
 }