+++ android/src/main/kotlin/com/maktubcompany/easy_blue_printer/plugin/domain/usecase/CommitPrintUseCase.kt
@@ -1,0 +1,7 @@
+package com.maktubcompany.easy_blue_printer.plugin.domain.usecase
+
+import com.maktubcompany.easy_blue_printer.plugin.domain.repository.BluetoothRepository
+
+class CommitPrintUseCase(private val repository: BluetoothRepository) {
+    fun execute(): Boolean = repository.commitPrint()
+}