+++ ios/Classes/plugin/domain/usecases/CommitPrintUseCase.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public class CommitPrintUseCase {
+    private let repository: BluetoothRepository
+
+    public init(repository: BluetoothRepository) {
+        self.repository = repository
+    }
+
+    public func execute() -> Bool {
+        return repository.commitPrint()
+    }
+}